requirements.txt
streamlit==1.33.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
python-dotenv==1.0.0
plotly==5.18.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyjwt==2.8.0
bcrypt==4.1.2